Extract HLS playback and controls-visibility hooks from VideoPlayer
VideoPlayer (489 lines) becomes a ~110-line composition. The playback
lifecycle (hls.js/native attachment, resume restoration, seek/stall
recovery) moves wholesale to useHlsPlayer; the idle-hide timer to
useControlsVisibility; the quality dropdown to player/QualityMenu.
Playback test suite passes unchanged.
